+++ packages/terra-functional-testing/CHANGELOG.md
@@ -6,6 +6,9 @@
   * Added authentication for accessing screenshots from the remote site.
   * Added `useRemoteReferenceScreenshots` cli option for downloading reference screenshots from a remote site for screenshot comparisons.
 
+* Fixed
+  * Fixed deleting the `diff`, `error`, and `latest` screenshot directories at the beginning of each test run.
+
 ## 2.7.0 - (February 11, 2022)
 
 * Changed